% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{allocation}
\alias{allocation}
\title{Largest Remainder Method}
\usage{
allocation(x, m, method, extra = NULL)
}
\arguments{
\item{x}{a data.frame or tibble with two columns. The first and second column represents names of parties and numbers of votes, respectively.}

\item{m}{a numeric value of district magnitude.}

\item{method}{a character string giving a method for computing seat allocation. See "Allocation methods" section.}

\item{extra}{a numeric vector of length \code{m}. If \code{method} is \code{"custom"}, this parameter is mandatory. Otherwise, it is ignored.}
}
\value{
A tibble object with two columns.
}
\description{
Largest Remainder Method
}
\details{
Custom methods (\code{"custom divisor"} and \code{"custom quota"}) require \code{extra} parameter. In the case of \code{"custom divisor"} (\code{"cd"}), a numeric vector of the same length as the magnitude (\code{m}) must be specified. If the magnitude (\code{m}) is a vector of length 2 or longer, the \code{extra} must have a length equal to \code{max(m)}. For example, if \code{m = c(5, 7, 4)}, then the length of the extra must be 7.

In the case of \code{"custom quota"}, \code{extra} must be a vector of the same length as \code{m}. For example, if \code{m = c(5, 7, 4)}, \code{extra} must also be a numeric vector of length 3. If length of \code{extra} is 1, the same quota is applied to all blocks.
}
\section{Allocation methods}{

\itemize{
\item Highest average method
\itemize{
\item \code{"d'hondt"} (\code{"dt"}) or \code{"jefferson"}: D'hondt / Jefferson method
\item \code{"adams"} or \code{"cambridge"}: Adams / Cambridge method
\item \code{"sainte-lague"} (\code{"sl"}) or \code{"webster"}: Sainte-Laguë / Webster method
\item \code{"modified sainte-lague"} (\code{"msl"}): Modified Sainte-Laguë method
\item \code{"danish"}: Danish method
\item \code{"imperiali"}: Imperiali method
\item \code{"hungtington-hill"} (\code{"hh"}): Huntington-Hill method
\item \code{"dean"}: Dean method
\item \code{"ad"}: alpha-divergence. \code{extra} is mandatory.
\item \code{"custom divisor"} (\code{"cd"}): Custom divisor. \code{extra} is mandatory.
}
\item Largest reminder method
\itemize{
\item \code{"hare-niemeyer"} (\code{"hn"} or \code{"hare"}): Hare-Niemeyer quota
\item \code{"droop"}: Droop quota
\item \code{"hagenbach-bischoff"} (\code{"hb"}): Hagenbach-Bischoff quota
\item \code{"imperiali quota"} (\code{"iq"}): Imperiali quota
\item \code{"custom quota"} (\code{"cq"}): Custom quota. \code{extra} is mandatory.
}
}
}

\examples{
sample_data <- data.frame(Party = c("Party A", "Party B", "Party C", "Party D"),
                          Votes = c(53000, 25000, 16600, 5400))

sample_data

# Magnitude = 10, Hare-Niemeyer quota
allocation(x = sample_data, m = 10, method = "hare")

# Magnitude = 10, D'hondt method
allocation(x = sample_data, m = 10, method = "dt")

# Custom divisor: 1.4, 3, 5, 7, ... (identical to Modified Sainte-Laguë method)
allocation(x = sample_data, m = 10, method = "custom",
           extra = c(1.4, seq(3, 19, by = 2)))
}
