% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_prcalc.R
\name{as_prcalc}
\alias{as_prcalc}
\title{Transformation from \code{data.frame} to \code{prcalc_df}}
\usage{
as_prcalc(
  x,
  region = NULL,
  district = NULL,
  population = NULL,
  magnitude = NULL,
  ...
)
}
\arguments{
\item{x}{a \code{data.frame} or \code{tibble} object.}

\item{region}{a character. A column name of region or state.}

\item{district}{a character. A column name of district.}

\item{population}{a character. A column name of population.}

\item{magnitude}{a character. A column name of magnitude. If \code{NULL}, magnitudes of all districts are replaced by \code{1}.}

\item{...}{ignored}
}
\value{
a \code{prcalc} object.
}
\description{
Transformation from \code{data.frame} to \code{prcalc_df}
}
\details{
an object \code{x} must include three columns---region(state) name, district name, and population (number of electorates). A column of magnitude is optional.
}
\examples{
library(PRcalc)
data(au_district_2010)

obj <- as_prcalc(au_district_2010,
                 region = "region",
                 district = "district",
                 population = "electorates",
                 magnitude = "magnitude")
obj
}
