% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{br_district_2010}
\alias{br_district_2010}
\title{List of electoral district in Brazil (2010)}
\format{
A data frame with 27 rows and 3 variables:
\describe{
\item{ district }{ names of state }
\item{ population }{ number of population }
\item{ magnitude }{ electoral magnitude }
}
}
\usage{
data("br_district_2010")
}
\description{
List of electoral district in Brazil (2010)
}
\keyword{datasets}
