% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\alias{compare.PRcalc}
\title{\code{PRcalc}オブジェクトの比較}
\usage{
compare(x, ...)

\method{compare}{PRcalc}(x, y, type = "seat", ...)
}
\arguments{
\item{x}{\code{PRcalc}オブジェクト}

\item{...}{Ignored}

\item{y}{\code{PRcalc}オブジェクト}

\item{type}{比較の基準; \code{"seat"} (議席数; 既定値)、または\code{"index"} (各種指標)}
}
\value{
A \code{PRcalc_compare} object
}
\description{
\code{PRcalc}オブジェクトの比較
}
