% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\alias{decompose.prcalc}
\title{Decompsition}
\usage{
decompose(x, ...)

\method{decompose}{prcalc}(x, alpha = 2, ...)
}
\arguments{
\item{x}{a \code{prcalc} object.}

\item{...}{ignored}

\item{alpha}{Default is \code{2}.}
}
\value{
a \code{list} object.
}
\description{
Decompsition
}
\examples{
data(jp_upper_2019)

pr_obj <- prcalc(jp_upper_2019, m = 50, method = "dt")

index(pr_obj)
}
\references{
\itemize{
\item{Laakso, Markku and Rein Taagepera. 1979. ""Effective" Number of Parties: A Measure with Application to West Europe". Comparative Political Studies. 12 (1): 3–27.}
}
}
