% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose3.R
\name{decompose3}
\alias{decompose3}
\alias{decompose3.prcalc}
\title{Decomposition using a \code{prcalc} object with three-step.}
\usage{
decompose3(x, ...)

\method{decompose3}{prcalc}(x, special = NULL, as_disprop = FALSE, alpha = 0, ...)
}
\arguments{
\item{x}{a \code{prcalc} object.}

\item{...}{ignored}

\item{as_disprop}{If \code{TRUE}, \code{alpha} must be larger than \code{0}. Default is \code{FALSE}.}

\item{alpha}{Default is \code{0}.}

\item{spcial}{a character. names of special district.}
}
\value{
a \code{prcalc_decomposition} object.
}
\description{
Decomposition using a \code{prcalc} object with three-step.
}
\examples{
print("Hello world!")
}
\references{
\itemize{
\item{Yuta, Kamahara, Junichiro Wada, and Yuko Kasuya. 2021. "Malapportionment in space and time: Decompose it!" \emph{Electoral Studies}. 71: 102301.}
}
}
\seealso{
\code{\link{decompose}}
}
