% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index.PRcalc}
\alias{index.PRcalc}
\title{各種指標の計算}
\usage{
\method{index}{PRcalc}(x, ...)
}
\arguments{
\item{x}{\code{PRcalc}オブジェクト}

\item{...}{Ignored}
}
\value{
A \code{PRcalc_index} object.
}
\description{
各種指標の計算
}
\references{
Laakso, Markku and Rein Taagepera. 1979. ""Effective" Number of Parties: A Measure with Application to West Europe". Comparative Political Studies. 12 (1): 3–27.
Gallagher, Michael. 1991. "Proportionality, Disproportionality and Electoral Systems". Electoral Studies. 10: 33–51.
}
