% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{it_distirct_2006}
\alias{it_distirct_2006}
\title{List of electoral district in Italy (2006)}
\format{
A data frame with 27 rows and 3 variables:
\describe{
\item{ region }{ names of region }
\item{ district }{ names of district }
\item{ population }{ number of population }
\item{ mag }{ electoral magnitude }
}
}
\usage{
data("it_distirct_2006")
}
\description{
List of electoral district in Italy (2006)
}
\section{Special region/district}{

There are four special region---"EUROPA", "AMERICA MERIDIONALE", "AMERICA SETTENTRIONALE E CENTRALE", and "AFRICA ASIA OCEANIA ANTARTIDE".
}

\keyword{datasets}
