% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{it_lower_2006}
\alias{it_lower_2006}
\title{Result in Italy lower house election (2006)}
\format{
A data frame with 46 rows and 4 variables:
\describe{
\item{ coalition }{ name of coalition }
\item{ party }{ name of party }
\item{ votes }{ number of votes }
\item{ seats }{ number of seats }
}
}
\usage{
data("it_lower_2006")
}
\description{
Result in Italy lower house election (2006)
}
\keyword{datasets}
