% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jp_census_2015}
\alias{jp_census_2015}
\title{都道府県別人口 (2015年)}
\format{
A data frame with 47 rows and 2 variables:
\describe{
\item{ Pref }{ 都道府県名; character }
\item{ Pop }{ 人口; numeric }
}
}
\source{
国勢調査
}
\usage{
data(jp_census_2015)
}
\description{
47都道府県の人口 (2015年)
}
\keyword{datasets}
