% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/--na.R
\docType{data}
\name{jp_lower_2021}
\alias{jp_lower_2021}
\title{衆議院議員総選挙比例区得票数 (2021年)}
\format{
A data frame with 47 rows and 12 variables:
\describe{
\item{ Party }{ 政党名; character }
\item{ Hokkaido }{ 北海道ブロックの得票数; numeric }
\item{ Tohoku }{ 東北ブロックの得票数; numeric }
\item{ Kitakanto }{ 北関東ブロックの得票数; numeric }
\item{ Minamikanto }{ 南関東ブロックの得票数; numeric }
\item{ Tokyo }{ 東京ブロックの得票数; numeric }
\item{ Hokuriku }{ 北陸信越ブロックの得票数; numeric }
\item{ Tokai }{ 東海ブロックの得票数; numeric }
\item{ Kinki }{ 近畿ブロックの得票数; numeric }
\item{ Chugoku }{ 中国ブロックの得票数; numeric }
\item{ Shikoku }{ 四国ブロックの得票数; numeric }
\item{ Kyushu }{ 九州ブロックの得票数; numeric }
}
}
\source{
総務省
}
\usage{
data(jp_lower_2021)
}
\description{
衆議院議員総選挙比例区の得票数 (2021年)
}
\keyword{datasets}
