% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jp_lower_2021_result}
\alias{jp_lower_2021_result}
\title{Result in Japanese lowerhouse election (2011)}
\format{
A data frame with 103 rows and 4 variables:
\describe{
\item{ Party }{ name of party }
\item{ Pref }{ name of prefecture }
\item{ Votes }{ number of votes }
\item{ Seats }{ number of seats }
}
}
\usage{
data(jp_lower_2021_result)
}
\description{
Result in Japanese lowerhouse election (2011)
}
\keyword{datasets}
