% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jp_upper_2019}
\alias{jp_upper_2019}
\title{参議院議員通常選挙比例区得票数 (2019年)}
\format{
A data frame with 47 rows and 2 variables:
\describe{
\item{ Party }{ 政党名; character }
\item{ Vote }{ 各政党の得票数; numeric }
}
}
\source{
総務省
}
\usage{
data(jp_upper_2019)
}
\description{
参議院議員通常選挙比例区の得票数 (2019年)
}
\keyword{datasets}
