% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/--na.R
\docType{data}
\name{kr_lower_2016}
\alias{kr_lower_2016}
\title{韓国国会議員総選挙比例区得票数 (2016年)}
\format{
A data frame with 47 rows and 2 variables:
\describe{
\item{ Party }{ 政党名; character }
\item{ Vote }{ 各政党の得票数; numeric }
}
}
\source{
韓国中央選挙管理委員会
}
\usage{
data(kr_lower_2016)
}
\description{
韓国国会議員総選挙比例区の得票数 (2016年)
}
\keyword{datasets}
