% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nz_district}
\alias{nz_district}
\title{List of electoral district in New Zealand}
\format{
A data frame with 1800 rows and 5 variables:
\describe{
\item{ year }{ election year }
\item{ region }{ names of region. \emph{Maori} is a special region. }
\item{ district }{ names of districts }
\item{ electorates }{ number of electorates }
\item{ magnitude }{ electoral magnitude }
}
}
\usage{
data(nz_district)
}
\description{
List of electoral district in New Zealand
}
\keyword{datasets}
