% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.prcalc_compare}
\alias{plot.prcalc_compare}
\title{Visualization of a \code{prcalc_compare} object.}
\usage{
\method{plot}{prcalc_compare}(
  x,
  bar_width = 0.75,
  facet = FALSE,
  free_y = TRUE,
  font_size = 12,
  angle = 0,
  facet_col = 4,
  title = NULL,
  caption = NULL,
  legend_pos = "bottom",
  ...
)
}
\arguments{
\item{x}{a \code{prcalc_compare} object.}

\item{bar_width}{Default is \code{0.75}.}

\item{facet}{政党ごとにfacet分割. Default is \code{FALSE}.}

\item{free_y}{Default is \code{TRUE}.}

\item{font_size}{a font size.}

\item{angle}{an angle of x-ticks label (0 to 90). Defualt is 0.}

\item{facet_col}{a number of columns of facets.}

\item{title}{a title of plot.}

\item{caption}{a caption of plot.}

\item{legend_pos}{a position of legend. Default is \code{"bottom"}}

\item{...}{Ignored}
}
\value{
A \code{ggplot} object.
}
\description{
Visualization of a \code{prcalc_compare} object.
}
\examples{
data(jp_upper_2019)

pr_obj1 <- prcalc(jp_upper_2019, m = 50, method = "dt")
pr_obj2 <- prcalc(jp_upper_2019, m = 50, method = "dt", threshold = 0.025)
pr_obj3 <- prcalc(jp_upper_2019, m = 50, method = "dt", threshold = 0.05)

list("t = 0\%"   = pr_obj1,
     "t = 2.5\%" = pr_obj2,
     "t = 5\%"   = pr_obj3) |>
  compare() |>
  plot()
}
