% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.prcalc_decomposition_compare}
\alias{plot.prcalc_decomposition_compare}
\title{Visualization of a \code{prcalc_decomposition_compare} object.}
\usage{
\method{plot}{prcalc_decomposition_compare}(
  x,
  percentage = FALSE,
  facet = FALSE,
  bar_width = 0.75,
  value_type = "all",
  value_angle = 0,
  value_size = 3,
  x_angle = 0,
  font_size = 12,
  border_width = 0,
  border_color = "black",
  digits = 3,
  ...
)
}
\arguments{
\item{x}{a \code{prcalc_decomposition_compare} object.}

\item{percentage}{If \code{TRUE}, the percentage of each element is displayed. If \code{facet = TRUE}, it is ignored.}

\item{facet}{Default is \code{FALSE}.}

\item{bar_width}{Default is \code{0.75}.}

\item{value_type}{\code{"all"}, \code{"total"}, or \code{"none"}. If \code{"total"}, only alpha-divergences are displayed. Default is \code{"all"}.}

\item{value_angle}{Angle of values. Default is \code{0}.}

\item{value_size}{Default is \code{3}. If \code{value_type == "none"}, it is ignored.}

\item{x_angle}{an angle of x-ticks label (\code{0} to \code{90}). Defualt is \code{0}.}

\item{font_size}{a font size. Default is \code{12}.}

\item{border_width}{width of border. Default is \code{0} (no border).}

\item{border_color}{color of border. Default is \code{"black"}.}

\item{digits}{Default is \code{3}.}

\item{...}{Ignored}
}
\value{
A \code{ggplot} object.
}
\description{
Visualization of a \code{prcalc_decomposition_compare} object.
}
\examples{
data(jp_lower_2021)

obj1 <- prcalc(jp_lower_2021,
               m = c(8, 13, 19, 22, 17, 11, 21, 28, 11, 6, 20),
               method = "dt")

obj2 <- prcalc(jp_lower_2021,
               m = c(8, 13, 19, 22, 17, 11, 21, 28, 11, 6, 20),
               method = "hare")

obj3 <- prcalc(jp_lower_2021,
               m = c(8, 13, 19, 22, 17, 11, 21, 28, 11, 6, 20),
               method = "msl")

obj4 <- prcalc(jp_lower_2021,
               m = c(8, 13, 19, 23, 19, 11, 20, 29, 10, 5, 19),
               method = "dt")

obj5 <- prcalc(jp_lower_2021,
               m = c(8, 13, 19, 23, 19, 11, 20, 29, 10, 5, 19),
               method = "msl")

compare(list("D'Hondt"     = decompose(obj1),
             "Hare"        = decompose(obj2),
             "Modified-SL" = decompose(obj3))) |>
  plot()

compare(list("Model 1" = decompose(obj1),
             "Model 2" = decompose(obj4),
             "Model 3" = decompose(obj5))) |>
  plot(facet = TRUE, value_type = "total", digits = 5)
}
