% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.prcalc_index}
\alias{plot.prcalc_index}
\title{Visualization of a \code{prcalc_index} object.}
\usage{
\method{plot}{prcalc_index}(
  x,
  index = NULL,
  style = "bar",
  bar_width = 0.75,
  point_size = 5,
  font_size = 12,
  ...
)
}
\arguments{
\item{x}{a \code{prcalc_index} object.}

\item{index}{a character vector. A subset of indices. If \code{NULL}, all indices are displayed. Default is \code{NULL}.}

\item{style}{Plot style. Lollipop (\verb{"lollipop}) or bar plot (\code{"bar"}). Default is \code{"bar"}.}

\item{bar_width}{Default is \code{0.75}.}

\item{point_size}{Default is \code{5}.}

\item{font_size}{a font size. Default is \code{12}.}

\item{...}{Ignored}
}
\value{
A \code{ggplot} object.
}
\description{
Visualization of a \code{prcalc_index} object.
}
\examples{
data(jp_upper_2019)

pr_obj <- prcalc(jp_upper_2019, m = 50, method = "dt")

plot(index(pr_obj))

plot(index(pr_obj), style = "lollipop",
     index = c("lh", "gallagher", "rae", "dhondt", "ad"))

}
