% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.prcalc_index_compare}
\alias{plot.prcalc_index_compare}
\title{Visualization of a \code{prcalc_index_compare} object.}
\usage{
\method{plot}{prcalc_index_compare}(
  x,
  index = NULL,
  style = "bar",
  bar_width = 0.75,
  point_size = 5,
  facet = FALSE,
  free_x = TRUE,
  font_size = 12,
  facet_col = 4,
  legend_pos = "bottom",
  ...
)
}
\arguments{
\item{x}{a \code{prcalc_index_compare} object.}

\item{index}{a character vector of disproportionality indices. If \code{NULL}, all indices are displayed.}

\item{style}{Plot style. Lollipop (\verb{"lollipop}) or bar plot (\code{"bar"}). Default is \code{"bar"}.}

\item{bar_width}{Default is \code{0.75}.}

\item{point_size}{Default is \code{5}.}

\item{facet}{政党ごとにfacet分割. Default is \code{FALSE}.}

\item{free_x}{Default is \code{TRUE}.}

\item{font_size}{a font size. Default is \code{12}.}

\item{facet_col}{a number of columns of facets. Default is \code{4}.}

\item{legend_pos}{a position of legend. Default is \code{"bottom"}}

\item{...}{Ignored}
}
\value{
A \code{ggplot} object.
}
\description{
Visualization of a \code{prcalc_index_compare} object.
}
\examples{
data(jp_upper_2019)

pr_obj1 <- prcalc(jp_upper_2019, m = 50, method = "dt")
pr_obj2 <- prcalc(jp_upper_2019, m = 50, method = "dt", threshold = 0.025)
pr_obj3 <- prcalc(jp_upper_2019, m = 50, method = "dt", threshold = 0.05)

list("t = 0\%"   = index(pr_obj1),
     "t = 2.5\%" = index(pr_obj2),
     "t = 5\%"   = index(pr_obj3)) |>
  compare() |>
  plot()

list("t = 0\%"   = index(pr_obj1),
     "t = 2.5\%" = index(pr_obj2),
     "t = 5\%"   = index(pr_obj3)) |>
  compare() |>
  plot(index = c("lh", "gallagher", "rae", "dhondt", "ad"))
}
