% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PRcalc_ex.R
\docType{package}
\name{PRcalc.ex}
\alias{PRcalc.ex}
\alias{PRcalc.ex-package}
\title{PRcalc.ex}
\usage{
PRcalc.ex(country)
}
\arguments{
\item{country}{\itemize {
 \item "korea": 2012 Korean General Election
 \item "japan": 2013 Japanese Upper House Election
 \item "austria": 2012 Austrian General election
}}
}
\description{
Sample of PRcalc for R
}
\examples{
#2012 Korean General Election
PR.calc("korea")
}
\author{
Jaehyun SONG
}

