% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcalc_block.R
\name{prcalc_block}
\alias{prcalc_block}
\title{Block PR}
\usage{
prcalc_block(votes, seat, method, threshold = 0)
}
\arguments{
\item{votes}{a data.frame or tibble with two columns. The first and second column represents names of parties and numbers of votes, respectively.}

\item{seat}{a numeric value of number of seats.}

\item{method}{a character string giving a method for computing seat allocation. This must be one of the strings "dt", "sl", "msl", "adams", "danish", "hh", "dean", or "imperiali".}

\item{threshold}{an optional numeric value of threshold. The}
}
\value{
A data.frame with two columns.
}
\description{
Block PR
}
\examples{
block_example <- data.frame(Party    = c("Party A", "Party B", "Party C", "Party D"),
                            Region_A = c(53000, 25000, 16600, 5400),
                            Region_B = c(12000, 34000, 5000, 10000),
                            Region_C = c(40000, 30000, 20000, 9000))

block_example

prcalc_block(votes       = block_example,
             seat        = c(10, 8, 11),
             method      = "dt",
             threshold   = 0)
}
