% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcalc_hae.R
\name{prcalc_hae}
\alias{prcalc_hae}
\title{Highest Average}
\usage{
prcalc_hae(
  votes,
  seat,
  method = c("dt", "sl", "msl", "adams", "danish", "hh", "dean", "imperiali"),
  threshold = 0,
  region_name = "National"
)
}
\arguments{
\item{votes}{a data.frame or tibble with two columns. The first and second column represents names of parties and numbers of votes, respectively.}

\item{seat}{a numeric value of number of seats.}

\item{method}{a character string giving a method for computing seat allocation. This must be one of the strings "dt", "sl", "msl", "adams", "danish", "hh", "dean", or "imperiali".}

\item{threshold}{an optional numeric value of threshold. The}

\item{region_name}{a character string.}
}
\value{
A data.frame with two columns.
}
\description{
Highest Average
}
\examples{
hae_example <- data.frame(Party = c("Party A", "Party B", "Party C", "Party D"),
                          Votes = c(53000, 25000, 16600, 5400))

hae_example

prcalc_hae(votes       = hae_example,
           seat        = 10,
           method      = "dt",
           threshold   = 0,
           region_name = "National")
}
