% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcalc_lrm.R
\name{prcalc_lrm}
\alias{prcalc_lrm}
\title{Largest Remainder Method}
\usage{
prcalc_lrm(
  votes,
  seat,
  method = c("hare", "droop", "imperialiQ"),
  threshold = 0,
  region_name
)
}
\arguments{
\item{votes}{a data.frame or tibble with two columns. The first and second column represents names of parties and numbers of votes, respectively.}

\item{seat}{a numeric value of number of seats.}

\item{method}{a character string giving a method for computing seat allocation. This must be one of the strings "hare", "droop", or "imperialiQ".}

\item{threshold}{an optional numeric value of threshold. The}

\item{region_name}{a character string.}
}
\value{
A data.frame with two columns.
}
\description{
Largest Remainder Method
}
\examples{
lrm_example <- data.frame(Party = c("Party A", "Party B", "Party C", "Party D"),
                          Votes = c(53000, 25000, 16600, 5400))

lrm_example

prcalc_lrm(votes       = lrm_example,
           seat        = 10,
           method      = "hare",
           threshold   = 0,
           region_name = "National")
}
