% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{print.PRcalc_compare}
\alias{print.PRcalc_compare}
\title{\code{PRcalc}オブジェクトの比較}
\usage{
\method{print}{PRcalc_compare}(x, digits = 3, ...)

\method{print}{PRcalc_compare}(x, digits = 3, ...)

\method{print}{PRcalc_compare}(x, digits = 3, ...)

\method{print}{PRcalc_compare}(x, digits = 3, ...)
}
\arguments{
\item{x}{\code{PRcalc_compare}オブジェクト}

\item{digits}{小数点の表示桁数}

\item{...}{Ignored}
}
\value{
A \code{PRcalc_compare} object

A \code{PRcalc_compare} object

A \code{PRcalc_compare} object

A \code{PRcalc_compare} object
}
\description{
\code{PRcalc}オブジェクトの比較

\code{PRcalc}オブジェクトの比較

\code{PRcalc}オブジェクトの比較

\code{PRcalc}オブジェクトの比較
}
\examples{
# アラバマのパラドックス
Election_Data3 <- data.frame(Party = c("Party A", "Party B", "Party C"),
                             Votes = c(5665, 3685, 1650))

PR_example3 <- PRcalc(Election_Data3, seats = 10, method = "hare")
PR_example4 <- PRcalc(Election_Data3, seats = 11, method = "hare")

compare(PR_example3, PR_example4)
# 議席割当方式の比較
PR_example5 <- PRcalc(Election_Data3, seats = 10, method = "dt")

compare(PR_example3, PR_example5)
compare1 <- compare(PR_example3, PR_example4, type = "index")
compare1
compare2 <- compare(PR_example3, PR_example5, type = "index")
print(compare2, digits = 1)
}
