% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{print.PRcalc_index}
\alias{print.PRcalc_index}
\title{Print \code{PRcalc_index} Object}
\usage{
\method{print}{PRcalc_index}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{PRcalc_index} object}

\item{digits}{a integer}

\item{...}{Ingnored}
}
\value{
none
}
\description{
Print \code{PRcalc_index} Object
}
\examples{
Election_Data1 <- data.frame(Party = c("Party A", "Party B", "Party C",
                                       "Party D", "Party E"),
                             Votes = c(778000, 714000, 331000, 299000, 230000))
Election_Data2 <- data.frame(Party   = c("Party A", "Party B", "Party C",
                                         "Party D", "Party E"),
                             Region1 = c(778000, 714000, 331000, 299000, 230000),
                             Region2 = c(1454000, 761000, 913000, 464000, 311000))

PR_example1 <- PRcalc(Election_Data1, seats = 8, method = "msl", threshold = 0)
PR_example2 <- PRcalc(Election_Data1, seats = 8, method = "msl", threshold = 0.1)
PR_example3 <- PRcalc(Election_Data2, seats = c(8, 13), method = "dt", threshold = 0)

PR_index1 <- index(PR_example1)
PR_index2 <- index(PR_example2)
PR_index3 <- index(PR_example3)

PR_index1
print(PR_index2, digits = 1)
PR_index3$gallagher # Gallagher非比例性指標のみ抽出
}
