% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prcalc.R
\name{read_prcalc}
\alias{read_prcalc}
\title{Transformation from \code{.csv} to \code{prcalc_df}}
\usage{
read_prcalc(
  x,
  region = NULL,
  district = NULL,
  population = NULL,
  magnitude = NULL,
  ...
)
}
\arguments{
\item{x}{a character. a path of \code{.csv} file.}

\item{region}{a character. A column name of region or state.}

\item{district}{a character. A column name of district.}

\item{population}{a character. A column name of population.}

\item{magnitude}{a character. A column name of magnitude. If \code{NULL}, magnitudes of all districts are replaced by \code{1}.}

\item{...}{ignored}
}
\value{
a \code{prcalc} object.
}
\description{
Transformation from \code{.csv} to \code{prcalc_df}
}
\details{
\code{.csv} file must include three columns---region(state) name, district name, and population (number of electorates). A column of magnitude is optional.
}
\examples{
\dontrun{
library(PRcalc)

my_df <- read_prcalc("data/my_file.csv")
my_df
}
}
