% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.prcalc}
\alias{summary.prcalc}
\title{Summrizing a \code{prcalc} object.}
\usage{
\method{summary}{prcalc}(object, prop = FALSE, use_gt = FALSE, digits = 3, ...)
}
\arguments{
\item{object}{a \code{prcalc} object.}

\item{prop}{If \code{TRUE}, voteshare and seatshare are displayed. Default is \code{FALSE}.}

\item{use_gt}{If \code{TRUE}, a table is rendered using \code{{gt}} package. Default is \code{FALSE}.}

\item{digits}{the number of decimal places. Default is 3.}

\item{...}{ignored.}
}
\description{
Summrizing a \code{prcalc} object.
}
\examples{
data(jp_upper_2019)
jp2019 <- prcalc(jp_upper_2019, m = 50, method = "dt")

summary(jp2019)

}
\seealso{
\code{\link{print.prcalc}}
}
