% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{us_census_2020}
\alias{us_census_2020}
\title{アメリカ州別人口 (2020年)}
\format{
A data frame with 47 rows and 2 variables:
\describe{
\item{ State }{ 州名; character }
\item{ Pop }{ 人口; numeric }
}
}
\source{
U.S. Census Bureau
}
\usage{
data(us_census_2020)
}
\description{
アメリカ州別人口 (2020年)
}
\keyword{datasets}
