% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_prcalc.R, R/read_prcalc.R
\name{as_prcalc}
\alias{as_prcalc}
\alias{read_prcalc}
\title{Transformation from \code{.csv} or \code{data.frame} (\code{tibble}) to \code{prcalc} object.}
\usage{
as_prcalc(x, l1, l2 = NULL, p, q = NULL, type = c("district", "party"), ...)

read_prcalc(x, l1, l2 = NULL, p, q = NULL, type = c("district", "party"), ...)
}
\arguments{
\item{x}{a path of \code{.csv} file (\code{read_prcalc()}), or name of \code{data.frame} or \code{tibble} object (\code{as_prcalc()}).}

\item{l1}{a character. \emph{Mandatory.} A column name of level 1 (region or state).}

\item{l2}{a character. \emph{Optional.} A column name of level 2 (party or district).}

\item{p}{a character. \emph{Mandatory.} A column name of population or number of electorates.}

\item{q}{a character. \emph{Optional.} A column name of magnitude or number of allocated seats. If \code{NULL}, replaced by \code{1}.}

\item{type}{a type of \code{l2}. \emph{Mandatory.} If \code{"party"}, vector of \code{l2} represents parties. If \code{"district"}, vector of \code{l2} represents electoral district. Default is \code{"district"}.}

\item{...}{ignored}
}
\value{
a \code{prcalc} object.
}
\description{
Transformation from \code{.csv} or \code{data.frame} (\code{tibble}) to \code{prcalc} object.
}
\details{
At least two columns of \code{x} are required---\code{l1} and \code{p}.
}
\examples{
data("jp_lower_2021_result")

as_prcalc(jp_lower_2021_result,
          l1   = "Pref",
          l2   = "Party",
          p    = "Votes",
          q    = "Seats",
          type = "party")

data("br_district_2010")
# Brazil has single block PR system, l2 can be omitted.
# If type = "party", raw names of districts are preserved.
br_district_2010 |>
  as_prcalc(l1   = "district",
            p    = "population",
            q    = "magnitude",
            type = "party")

data("au_district_2010")
# Because all of magnitude are 1 and type of l2 is district,
# you can omit q and type arguments.
as_prcalc(au_district_2010,
          l1   = "region",
          l2   = "district",
          p    = "electorates")
# You can import .csv format and transfrom into prcalc object directly.
\dontrun{
read_prcalc("data/my_file.csv",
            l1   = "region",
            l2   = "district",
            p    = "electorates",
            q    = "magnitude",
            type = "district")
}
}
