% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{au_district_2010}
\alias{au_district_2010}
\title{List of electoral district in Australia (2010)}
\format{
A data frame with 150 rows and 4 variables:
\describe{
\item{ region }{ name of region }
\item{ district }{ name of district }
\item{ electorates }{ number of electorates }
\item{ magnitude }{ magnitude }
}
}
\usage{
data(au_district_2010)
}
\description{
List of electoral district in Australia (2010)
}
\keyword{datasets}
