% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare.PRcalc}
\alias{compare.PRcalc}
\title{\code{PRcalc}オブジェクトの比較}
\usage{
\method{compare}{PRcalc}(x, y, type = "seat", digits = 3, ...)

\method{compare}{PRcalc}(x, y, type = "seat", digits = 3, ...)

\method{compare}{PRcalc}(x, y, type = "seat", digits = 3, ...)

\method{compare}{PRcalc}(x, y, type = "seat", digits = 3, ...)
}
\arguments{
\item{x}{\code{PRcalc}オブジェクト}

\item{y}{\code{PRcalc}オブジェクト}

\item{type}{比較の基準; \code{"seat"} (議席数; 既定値)、または\code{"index"} (各種指標)}

\item{digits}{小数点の表示桁数 (\code{type = "index"}の場合のみ有効)}

\item{...}{Ignored}
}
\value{
none

none

none

none
}
\description{
\code{PRcalc}オブジェクトの比較

\code{PRcalc}オブジェクトの比較

\code{PRcalc}オブジェクトの比較

\code{PRcalc}オブジェクトの比較
}
\examples{
# アラバマのパラドックス
Election_Data3 <- data.frame(Party = c("Party A", "Party B", "Party C"),
                             Votes = c(5665, 3685, 1650))

PR_example3 <- PRcalc(Election_Data3, seats = 10, method = "hare")
PR_example4 <- PRcalc(Election_Data3, seats = 11, method = "hare")

compare(PR_example3, PR_example4)
# 議席割当方式の比較
PR_example5 <- PRcalc(Election_Data3, seats = 10, method = "dt")

compare(PR_example3, PR_example5)
compare(PR_example3, PR_example4, type = "index")
compare(PR_example3, PR_example5, type = "index", digits = 1)
}
