% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\alias{compare.list}
\title{Comparison multiple \code{prcalc}, \code{prcalc_index}, or \code{prcalc_decomposition} objects.}
\usage{
compare(x, ...)

\method{compare}{list}(x, prop = FALSE, ...)
}
\arguments{
\item{x}{a list. All elements in the list must be \code{prcalc}, \code{prcalc_index}, or \code{prcalc_decomposition} objects.}

\item{...}{Ignored}

\item{prop}{Default is \code{FALSE}.}
}
\value{
A \code{prcalc_compare}, \code{prcalc_index_compare}, or \code{prcalc_decomposition_compare} object
}
\description{
Comparison multiple \code{prcalc}, \code{prcalc_index}, or \code{prcalc_decomposition} objects.
}
\examples{
data(jp_upper_2019)

pr_obj1 <- prcalc(jp_upper_2019, m = 50, method = "dt")
pr_obj2 <- prcalc(jp_upper_2019, m = 50, method = "dt", threshold = 0.025)
pr_obj3 <- prcalc(jp_upper_2019, m = 50, method = "dt", threshold = 0.05)

compare(list("t = 0\%" = pr_obj1, "t = 2.5\%" = pr_obj2, "t = 5\%" = pr_obj3))

list("t = 0\%"   = index(pr_obj1),
     "t = 2.5\%" = index(pr_obj2),
     "t = 5\%"   = index(pr_obj3)) |>
  compare()
}
