% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\alias{decompose.prcalc}
\title{Decomposition using a \code{prcalc} object.}
\usage{
decompose(x, ...)

\method{decompose}{prcalc}(x, as_disprop = TRUE, alpha = 2, ...)
}
\arguments{
\item{x}{a \code{prcalc} or \code{prcalc_df} object.}

\item{...}{ignored}

\item{as_disprop}{If \code{TRUE}, \code{alpha} must be larger than \code{0}. Default is \code{TRUE}.}

\item{alpha}{Default is \code{2} and it must be larger than \code{0}.}
}
\value{
a \code{prcalc_decomposition} object.
}
\description{
Decomposition using a \code{prcalc} object.
}
\examples{
# Using a prcalc object
data(jp_lower_2021)

obj <- prcalc(jp_lower_2021,
              m = c(8, 14, 20, 21, 17, 11, 21, 30, 11, 6, 21),
              method = "hare")

decompose(obj)
}
\references{
\itemize{
\item{Yuta, Kamahara. "The Desired Political Entropy as the Measure of Unequal Representation: Disproportionality and Malapportionment". Working paper.}
}
}
