% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\alias{decompose.prcalc}
\title{Decomposition using a \code{prcalc} object.}
\usage{
decompose(x, as_disprop = TRUE, alpha = 2, special = NULL, ...)

\method{decompose}{prcalc}(x, as_disprop = TRUE, alpha = 2, special = NULL, ...)
}
\arguments{
\item{x}{a \code{prcalc} object.}

\item{as_disprop}{If \code{TRUE}, \code{alpha} must be larger than \code{0}. Default is \code{TRUE}.}

\item{alpha}{Default is \code{2} and it must be larger than \code{0}.}

\item{special}{a character. names of special district. If it is defined, three-step decomposition is conducted.}

\item{...}{ignored}
}
\value{
a \code{prcalc_decomposition} object.
}
\description{
Decomposition using a \code{prcalc} object.
}
\examples{
# two-step decomposition (example 1)
data("jp_lower_2021")

obj <- prcalc(jp_lower_2021,
              m = c(8, 14, 20, 21, 17, 11, 21, 30, 11, 6, 21),
              method = "hare")

decompose(obj)

# two-step decomposition (example 2)
data("au_district_2010")

au_data <- au_district_2010 |>
  as_prcalc(region     = "region",
            district   = "district",
            population = "electorates",
            magnitude  = "magnitude")

decompose(au_data, alpha = 0, as_disprop = FALSE)

# three-step decomposition
data("nz_district")

nz_district |>
  dplyr::filter(year == 2011) |>
  as_prcalc(region     = "region",
            district   = "district",
            population = "electorates",
            magnitude  = "magnitude") |>
  decompose(alpha = 0, as_disprop = FALSE, special = "Maori")
}
\references{
\itemize{
\item{Yuta, Kamahara. "The Desired Political Entropy as the Measure of Unequal Representation: Disproportionality and Malapportionment". Working paper.}
}
}
\seealso{
\code{\link{decompose2}}, \code{\link{decompose3}}
}
