% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\alias{decompose.prcalc}
\title{Decompsition}
\usage{
decompose(x, ...)

\method{decompose}{prcalc}(x, alpha = 2, ...)
}
\arguments{
\item{x}{a \code{prcalc} object.}

\item{...}{ignored}

\item{alpha}{Default is \code{2}.}
}
\value{
a \code{prcalc_decomposition} object.
}
\description{
Decompsition
}
\examples{
data(jp_lower_2021)

obj <- prcalc(jp_lower_2021[1:9, ],
              m = c(8, 14, 20, 21, 17, 11, 21, 30, 11, 6, 21),
              method = "hare")

decompose(obj)
}
\references{
\itemize{
\item{Yuta, Kamahara. "The Desired Political Entropy as the Measure of Unequal Representation: Disproportionality and Malapportionment". Working paper.}
}
}
