% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose3.R
\name{decompose3}
\alias{decompose3}
\alias{decompose3.prcalc}
\title{Decomposition using a \code{prcalc} object with three-step.}
\usage{
decompose3(x, ...)

\method{decompose3}{prcalc}(x, special = NULL, alpha = 2, ...)
}
\arguments{
\item{x}{a \code{prcalc} object.}

\item{...}{ignored}

\item{special}{a character. names of special district.}

\item{alpha}{Default is \code{2}.}
}
\value{
a \code{prcalc_decomposition} object.
}
\description{
Decomposition using a \code{prcalc} object with three-step.
}
\examples{
data(nz_district)

nz_1949 <- nz_district |>
  subset(year == 1949) |>
  as_prcalc(l1   = "region",
            l2   = "district",
            p    = "electorates",
            q    = "magnitude",
            type = "district")

decompose3(nz_1949, special = "Maori")
}
\references{
\itemize{
\item{Yuta, Kamahara, Junichiro Wada, and Yuko Kasuya. 2021. "Malapportionment in space and time: Decompose it!" \emph{Electoral Studies}. 71: 102301.}
}
}
\seealso{
\code{\link{decompose}}, \code{\link{decompose2}}

\code{\link{decompose}}
}
