% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\alias{index.prcalc}
\title{各種指標の計算}
\usage{
index(x, ...)

\method{index}{prcalc}(x, k = 2, eta = 2, alpha = 2, ...)
}
\arguments{
\item{x}{a \code{prcalc} object.}

\item{...}{ignored}

\item{k}{a parameter for Generalized Gallagher index. Default is \code{2}.}

\item{eta}{a parameter for Atkinson (Gini) index. Default is \code{2}.}

\item{alpha}{a parameter for Generalized Entropy and alpha-divergence. Default is \code{2}.}
}
\value{
a \code{prcalc_index} object.
}
\description{
各種指標の計算
}
\examples{
data(jp_upper_2019)

pr_obj <- prcalc(jp_upper_2019, m = 50, method = "dt")

index(pr_obj)
}
\references{
\itemize{
\item{Laakso, Markku and Rein Taagepera. 1979. ""Effective" Number of Parties: A Measure with Application to West Europe". Comparative Political Studies. 12 (1): 3–27.}
\item{Gallagher, Michael. 1991. "Proportionality, Disproportionality and Electoral Systems". Electoral Studies. 10: 33–51.}
}
}
