% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\alias{index.prcalc}
\title{Calculate the disproportionality indices.}
\usage{
index(x, ...)

\method{index}{prcalc}(x, as_disprop = TRUE, k = 2, eta = 2, alpha = 2, omit_zero = TRUE, ...)
}
\arguments{
\item{x}{a \code{prcalc} object.}

\item{...}{ignored}

\item{as_disprop}{Measuring disporportionality? If \code{TRUE} (default), \code{alpha} must be larger than \code{0}. For measuring malapportionment, \code{as_disprop = FALSE}.}

\item{k}{a parameter for Generalized Gallagher index. Default is \code{2}.}

\item{eta}{a parameter for Atkinson index. Default is \code{2}.}

\item{alpha}{a parameter for alpha-divergence. \code{alpha} must be larger than 0. Default is \code{2}.}

\item{omit_zero}{If \code{TRUE}, parties with 0 votes and 0 seats are ignored. Default is \code{TRUE}.}
}
\value{
a \code{prcalc_index} object.
}
\description{
Calculate the disproportionality indices.
}
\examples{
data(jp_upper_2019)

obj <- prcalc(jp_upper_2019, m = 50, method = "dt")

obj_index <- index(obj)
obj_index

obj_index["gallagher"] # Extract Gallagher index
}
\references{
\itemize{
\item{Laakso, Markku and Rein Taagepera. 1979. ""Effective" Number of Parties: A Measure with Application to West Europe". Comparative Political Studies. 12 (1): 3–27.}
\item{Gallagher, Michael. 1991. "Proportionality, Disproportionality and Electoral Systems". Electoral Studies. 10: 33–51.}
}
}
