% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.prcalc}
\alias{plot.prcalc}
\title{Visualization of a \code{prcalc} object.}
\usage{
\method{plot}{prcalc}(
  x,
  type = "both",
  prop = TRUE,
  show_total = TRUE,
  by = "block",
  subset_b = NULL,
  subset_p = NULL,
  free_y = FALSE,
  font_size = 12,
  angle = 0,
  facet_col = 4,
  legend_pos = "bottom",
  xlab = "Parties",
  ylab = "Distribution",
  title = NULL,
  caption = NULL,
  ...
)
}
\arguments{
\item{x}{a \code{prcalc} object.}

\item{type}{a plot of votes (\code{"raw"}), allocation (\code{"dist"}), or both (\code{"both"}). Default is \code{"both"}.}

\item{prop}{割合で表示。If \code{type} is \code{"both"}, it is ignored. Default is \code{TRUE}.}

\item{show_total}{If \code{TRUE}, a facet names "Total" is diplayed. Default is \code{TRUE}.}

\item{by}{Facet分割の単位（\code{"block"} or \code{"party"}）。Default is \code{"block"}}

\item{subset_b}{a character vector of block names.}

\item{subset_p}{a character vector of party names.}

\item{free_y}{If \code{FALSE}, y-axes are fixed over all facets. Default is \code{TRUE}.}

\item{font_size}{a font size.}

\item{angle}{an angle of x-ticks label (0 to 90). Defualt is 0.}

\item{facet_col}{a number of columns of facets.}

\item{legend_pos}{a position of legend. Default is \code{"bottom"}}

\item{xlab}{a label of x-axis.}

\item{ylab}{a label of y-axis.}

\item{title}{a title of plot.}

\item{caption}{a caption of plot.}

\item{...}{Ignored}
}
\value{
A \code{ggplot} object.
}
\description{
Visualization of a \code{prcalc} object.
}
\examples{
data(jp_upper_2019)
data(jp_lower_2021)

pr_obj1 <- prcalc(jp_upper_2019, m = 50, method = "dt")
plot(pr_obj1)

pr_obj2 <- prcalc(jp_lower_2021,
                  m = c(8, 14, 20, 21, 17, 11, 21, 30, 11, 6, 21),
                  method = "hare")
plot(pr_obj2, angle = 90)
plot(pr_obj2, subset_p = c("自民", "公明", "立憲", "維新", "国民"))
plot(pr_obj2, subset_b = c("Tokyo", "Kinki"), facet_col = 1)
plot(pr_obj2, by = "party",
     subset_p = c("自民", "公明", "立憲", "維新"),
     subset_b = c("Hokkaido", "Tohoku", "Tokyo", "Kinki"))

}
