% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.prcalc}
\alias{plot.prcalc}
\title{Visualization of a \code{prcalc} object.}
\usage{
\method{plot}{prcalc}(
  x,
  type = "both",
  prop = TRUE,
  show_total = TRUE,
  by = c("l1", "l2"),
  subset_l1 = NULL,
  subset_l2 = NULL,
  free_y = FALSE,
  font_size = 12,
  angle = 0,
  facet_col = 4,
  legend_pos = "bottom",
  ...
)
}
\arguments{
\item{x}{a \code{prcalc} object.}

\item{type}{a plot of votes (\code{"raw"}), allocation (\code{"dist"}), or both (\code{"both"}). Default is \code{"both"}.}

\item{prop}{Output at a percentage. If \code{type} is \code{"both"}, it is ignored. Default is \code{TRUE}.}

\item{show_total}{If \code{TRUE}, a facet names "Total" is diplayed. Default is \code{TRUE}.}

\item{by}{Facets by level 1 (\code{"l1"}) or level 2 (\code{"l2"}). Default is \code{"l1"}.}

\item{subset_l1}{a character vector. a subset of level 1.}

\item{subset_l2}{a character vector. a subset of level 2.}

\item{free_y}{If \code{FALSE}, y-axes are fixed over all facets. Default is \code{TRUE}.}

\item{font_size}{a font size.}

\item{angle}{an angle of x-ticks label (\code{0} to \code{90}). Defualt is \code{0}.}

\item{facet_col}{a number of columns of facets.}

\item{legend_pos}{a position of legend. Default is \code{"bottom"}}

\item{...}{Ignored}
}
\value{
A \code{ggplot} object.
}
\description{
Visualization of a \code{prcalc} object.
}
\examples{
data(jp_upper_2019)
data(jp_lower_2021)

pr_obj1 <- prcalc(jp_upper_2019, m = 50, method = "dt")
plot(pr_obj1)

pr_obj2 <- prcalc(jp_lower_2021,
                  m = c(8, 14, 20, 21, 17, 11, 21, 30, 11, 6, 21),
                  method = "hare")
plot(pr_obj2, angle = 90)
plot(pr_obj2, subset_l2 = c("自民", "公明", "立憲", "維新", "国民"))
plot(pr_obj2, subset_l1 = c("Tokyo", "Kinki"), facet_col = 1)
plot(pr_obj2, by = "l2",
     subset_l2 = c("自民", "公明", "立憲", "維新"),
     subset_l1 = c("Hokkaido", "Tohoku", "Tokyo", "Kinki"))

}
