% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prcalc.R
\name{plot.PRcalc}
\alias{plot.PRcalc}
\title{\code{PRcalc}オブジェクトの可視化}
\usage{
\method{plot}{PRcalc}(
  x,
  summary = FALSE,
  xlab = "Party",
  ylab = "\%",
  vs_lab = "Vote share",
  ss_lab = "Seat share",
  text_size = 12,
  text_angle = 0,
  ...
)

\method{plot}{PRcalc}(
  x,
  summary = FALSE,
  xlab = "Party",
  ylab = "\%",
  vs_lab = "Vote share",
  ss_lab = "Seat share",
  text_size = 12,
  text_angle = 0,
  ...
)
}
\arguments{
\item{x}{\code{PRcalc}オブジェクト}

\item{summary}{\code{TRUE}、または\code{FALSE} (既定値); \code{TRUE}の場合、ブロックごとのプロットが省略されます。}

\item{xlab}{横軸のタイトル}

\item{ylab}{縦軸のタイトル}

\item{vs_lab}{凡例ラベル (得票率)}

\item{ss_lab}{凡例ラベル (議席率)}

\item{text_size}{文字の大きさ; 既定値は12}

\item{text_angle}{横軸ラベルの角度; 既定値は0}

\item{...}{Ingnored}
}
\value{
A \code{ggplot} object.

A \code{ggplot} object.
}
\description{
\code{PRcalc}オブジェクトの可視化

\code{PRcalc}オブジェクトの可視化
}
\examples{
#| dev: "ragg_png"
#| fig.width: 7
#| out.width: 650
#| dpi: 300
Election_Data1 <- data.frame(Party = c("Party A", "Party B", "Party C",
                                       "Party D", "Party E"),
                             Votes = c(778000, 714000, 331000, 299000, 230000))

PR_example1 <- PRcalc(Election_Data1, seats = 8, method = "msl", threshold = 0)

plot(PR_example1) # PR_example1の可視化
#| dev: "ragg_png"
#| fig.width: 7
#| out.width: 650
#| dpi: 300
Election_Data2 <- data.frame(Party   = c("Party A", "Party B", "Party C",
                                         "Party D", "Party E"),
                             Region1 = c(778000, 714000, 331000, 299000, 230000),
                             Region2 = c(1454000, 761000, 913000, 464000, 311000))

PR_example2 <- PRcalc(Election_Data2, seats = c(8, 13), method = "dt", threshold = 0)

plot(PR_example2) # PR_example2の可視化
plot(PR_example2, text_angle = 25) # X軸ラベルの回転
plot(PR_example2, summary = TRUE) # ブロックごとの図を省略
}
