% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.prcalc_decomposition_compare}
\alias{plot.prcalc_decomposition_compare}
\title{Visualization of a \code{prcalc_decomposition_compare} object.}
\usage{
\method{plot}{prcalc_decomposition_compare}(
  x,
  facet = FALSE,
  bar_width = 0.75,
  value_type = "label",
  value_size = 3,
  font_size = 12,
  digits = 3,
  ...
)
}
\arguments{
\item{x}{a \code{prcalc_decomposition_compare} object.}

\item{facet}{Default is \code{FALSE}.}

\item{bar_width}{Default is \code{0.75}.}

\item{value_type}{\code{"label"}, \code{"text"}, or \code{"none"}. Default is \code{"label"}.}

\item{value_size}{Default is \code{3}.}

\item{font_size}{a font size. Default is \code{12}.}

\item{digits}{Default is \code{3}.}

\item{...}{Ignored}
}
\value{
A \code{ggplot} object.
}
\description{
Visualization of a \code{prcalc_decomposition_compare} object.
}
\examples{
data(jp_lower_2021)

obj1 <- prcalc(jp_lower_2021,
               m = c(8, 13, 19, 22, 17, 11, 21, 28, 11, 6, 20),
               method = "dt")

obj2 <- prcalc(jp_lower_2021,
               m = c(8, 13, 19, 22, 17, 11, 21, 28, 11, 6, 20),
               method = "hare")

obj3 <- prcalc(jp_lower_2021,
               m = c(8, 13, 19, 22, 17, 11, 21, 28, 11, 6, 20),
               method = "msl")

obj4 <- prcalc(jp_lower_2021,
               m = c(8, 13, 19, 23, 19, 11, 20, 29, 10, 5, 19),
               method = "dt")

obj5 <- prcalc(jp_lower_2021,
               m = c(8, 13, 19, 23, 19, 11, 20, 29, 10, 5, 19),
               method = "msl")

compare(list("D'Hondt"     = decompose(obj1),
             "Hare"        = decompose(obj2),
             "Modified-SL" = decompose(obj3))) |>
  plot()

compare(list("Model 1" = decompose(obj1),
             "Model 2" = decompose(obj4),
             "Model 3" = decompose(obj5))) |>
  plot(facet = TRUE, value_type = "text", digits = 5)
}
