% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcalc.R
\name{prcalc}
\alias{prcalc}
\title{Proportional Representation Calculator}
\usage{
prcalc(x, m, method, extra = NULL, threshold = 0)
}
\arguments{
\item{x}{a \code{data.frame} or \code{tibble} object.}

\item{m}{a \code{numeric} vector of magnitude.}

\item{method}{a character string giving a method for computing seat allocation. See "Allocation methods" section.}

\item{extra}{a numeric vector of length \code{m}. If \code{method} is \code{"custom"}, this parameter is mandatory. Otherwise, it is ignored.}

\item{threshold}{a threshold. It must be equal or larger than 0 (0\%) and smaller than 0.1 (10\%). a default is 0.}
}
\value{
a \code{prcalc} object
}
\description{
Proportional Representation Calculator
}
\examples{
data(jp_upper_2019)
data(jp_lower_2021)
prcalc(jp_upper_2019, m = 50, method = "dt")

japan_2021 <- prcalc(jp_lower_2021,
                     m = c(8, 14, 20, 21, 17, 11, 21, 30, 11, 6, 21),
                     method = "hare")
japan_2021
}
\seealso{
\code{\link{print.prcalc}}, \code{\link{summary.prcalc}}, \code{\link{plot.prcalc}}
}
