% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRcalc.R
\name{prcalc}
\alias{prcalc}
\title{Proportional Representation Calculator}
\usage{
prcalc(x, m, method, extra = NULL, threshold = 0)
}
\arguments{
\item{x}{a \code{data.frame} or \code{tibble} object.}

\item{m}{a numeric vector of district magnitude.}

\item{method}{a character string giving a method for computing seat allocation. See "Allocation methods" section.}

\item{extra}{a numeric vector of length \code{m}. If \code{method} is \code{"custom divisor"} (\code{"cd"}) or \code{"custom quota"} (\code{"cq"}), this parameter is mandatory. See "About custom methods" section.}

\item{threshold}{a threshold. It must be equal or larger than 0 (0\%) and smaller than 0.1 (10\%). a default is 0.}
}
\value{
a \code{prcalc} object
}
\description{
Proportional Representation Calculator
}
\section{Allocation methods}{

aaa
}

\section{About custom methods}{

Custom methods (\code{"custom divisor"} and \code{"custom quota"}) require \code{extra} parameter. In the case of \code{"custom divisor"} (\code{"cd"}), a numeric vector of the same length as the magnitude (\code{m}) must be specified. If the magnitude (\code{m}) is a vector of length 2 or longer, the \code{extra} must have a length equal to \code{max(m)}. For example, if \code{m = c(5, 7, 4)}, then the length of the extra must be 7.

In the case of \code{"custom quota"}, \code{extra} must be a vector of the same length as \code{m}. For example, if \code{m = c(5, 7, 4)}, \code{extra} must also be a numeric vector of length 3. If length of \code{extra} is 1, the same quota is applied to all blocks.
}

\examples{
data(jp_upper_2019)
data(jp_lower_2021)
prcalc(jp_upper_2019, m = 50, method = "dt")

japan_2021 <- prcalc(jp_lower_2021,
                     m = c(8, 14, 20, 21, 17, 11, 21, 30, 11, 6, 21),
                     method = "hare")
japan_2021
}
\seealso{
\code{\link{print.prcalc}}, \code{\link{summary.prcalc}}, \code{\link{plot.prcalc}}
}
