% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcalc.R
\name{PRcalc}
\alias{PRcalc}
\title{PR Calculation}
\usage{
PRcalc(x, seats, method, threshold = 0)
}
\arguments{
\item{x}{a \code{data.frame} or \code{tibble} object.}

\item{seats}{議席数; 長さ1以上の\code{numeric}型ベクトル}

\item{method}{議席割当方式;}

\item{threshold}{a numeric value}
}
\value{
A \code{list} type object.
}
\description{
PR Calculation
}
