% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PRcalc.R
\docType{package}
\name{PRcalc}
\alias{PRcalc}
\alias{PRcalc-package}
\title{PRcalc}
\usage{
PRcalc(nseat, vote, method
       threshold = 0, viewer = TRUE)
}
\arguments{
\item{nseat}{Numver of seats(numeric)}

\item{vote}{Votes(vector or data.frame)}

\item{method}{\itemize {
 \item dt: d'Hondt
 \item sl: Sainte-Laguë
 \item msl: Modified Sainte-Laguë
 \item denmark: Danish
 \item imperiali: Imperiali
 \item hh: Huntington-Hill
 \item hare: Hare
 \item droop: Droop
 \item hh: Huntington-Hill
}}

\item{threshold}{Threshold(numeric, 0~1)}

\item{viewr}{Showing the result}
}
\value{
List type
}
\description{
Proportional Representation Calculator for R
}
\examples{
#Number of seats=50, (PartyA 100 votes, PartyB 80 votes, PartyC 60 votes), Method = Hare
PR.calc(nseat = 50,
        vote = c(“Party A” = 100,
                 “Party B” = 80,
                 “Party C” = 60),
        method = “hare”)

#Number of seats = 54, ( 700 votes, 80 votes, 100 votes, 70 votes (without party names)), Method = d’Hondt, Threshold = 5\%
PR.calc(nseat = 54,
        vote = c(700, 80, 100, 70),
        method = “dt”, threshold = 0.05)

# Number of seats = 183, Using Austrian sample dataset, Method = d’Hondt, Threshold = 4\%
PR.calc(nseat = 183,
        vote = austrian.sample,
        method = “dt”, threshold = 0.04)
}
\author{
Jaehyun SONG
}

