% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.prcalc}
\alias{print.prcalc}
\title{Printing a \code{prcalc} object.}
\usage{
\method{print}{prcalc}(
  x,
  prop = FALSE,
  show_total = TRUE,
  use_gt = FALSE,
  digits = 3,
  head = Inf,
  ...
)
}
\arguments{
\item{x}{a \code{prcalc} object.}

\item{prop}{If \code{TRUE}, voteshare and seatshare are displayed. Default is \code{FALSE}.}

\item{show_total}{If \code{TRUE}, total vote (share) and seat (share) are displayed on the last column. Default is \code{TRUE}.}

\item{use_gt}{Use \code{gt} package? Default is \code{FALSE}.}

\item{digits}{the number of decimal places. Default is 3.}

\item{head}{For example, if \code{head = 6}, only six rows are displayed. Default is \code{Inf} (entire rows).}

\item{...}{ignored.}
}
\description{
Printing a \code{prcalc} object.
}
\examples{
data(jp_upper_2019)
jp2019 <- prcalc(jp_upper_2019, m = 50, method = "dt")

print(jp2019)

}
\seealso{
\code{\link{summary.prcalc}}
}
