% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcalc.R
\name{print.PRcalc}
\alias{print.PRcalc}
\title{Print \code{PRclac} Object}
\usage{
\method{print}{PRcalc}(x, prop = FALSE, digits = 3, ...)
}
\arguments{
\item{x}{\code{PRcalc} object}

\item{prop}{a logical value}

\item{digits}{a integer}

\item{...}{Ingnored}
}
\value{
none
}
\description{
Print \code{PRclac} Object
}
