% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.prcalc_compare}
\alias{print.prcalc_compare}
\title{Printing a \code{prcalc_compare} object.}
\usage{
\method{print}{prcalc_compare}(x, subset = NULL, prop = FALSE, use_gt = FALSE, digits = 3, ...)
}
\arguments{
\item{x}{a \code{prcalc_compare} object.}

\item{subset}{a}

\item{prop}{a}

\item{use_gt}{Use \code{gt} packages? Default is \code{FALSE}.}

\item{digits}{the number of decimal places. Default is 3.}

\item{...}{ignored.}
}
\description{
Printing a \code{prcalc_compare} object.
}
\examples{
data(jp_upper_2019)

pr_obj1 <- prcalc(jp_upper_2019, m = 50, method = "dt")
pr_obj2 <- prcalc(jp_upper_2019, m = 50, method = "dt", threshold = 0.025)
pr_obj3 <- prcalc(jp_upper_2019, m = 50, method = "dt", threshold = 0.05)

compare(list("t = 0\%" = pr_obj1, "t = 2.5\%" = pr_obj2, "t = 5\%" = pr_obj3))

list("t = 0\%"   = pr_obj1,
     "t = 2.5\%" = pr_obj2,
     "t = 5\%"   = pr_obj3) |>
  compare() |>
  print()

list("t = 0\%"   = pr_obj1,
     "t = 2.5\%" = pr_obj2,
     "t = 5\%"   = pr_obj3) |>
  compare() |>
  print(subset = c("自民", "公明", "立憲", "共産", "維新"))

}
