% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.prcalc_index}
\alias{print.prcalc_index}
\title{Printing a \code{prcalc_index} object.}
\usage{
\method{print}{prcalc_index}(x, subset = NULL, hide_id = FALSE, use_gt = FALSE, digits = 3, ...)
}
\arguments{
\item{x}{a \code{prcalc_index} object.}

\item{subset}{a}

\item{hide_id}{a}

\item{use_gt}{a}

\item{digits}{the number of decimal places. Default is 3.}

\item{...}{ignored.}
}
\description{
Printing a \code{prcalc_index} object.
}
\examples{
data(jp_upper_2019)

pr_obj <- prcalc(jp_upper_2019, m = 50, method = "dt")

index(pr_obj)

index(pr_obj) |>
  print(subset = c("lh", "gallagher", "rae", "dhondt", "ad"))
}
