% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prcalc.R
\name{read_prcalc}
\alias{read_prcalc}
\title{Transformation from \code{.csv} to \code{prcalc}}
\usage{
read_prcalc(
  x,
  region = "region",
  district = "district",
  population = "electorates",
  magnitude = "magnitude",
  distinct_name = FALSE,
  ...
)
}
\arguments{
\item{x}{a character. a path of \code{.csv} file.}

\item{region}{a character (mandatory). A column name of region or state.}

\item{district}{a character (mandatory). A column name of district.}

\item{population}{a character (mandatory). A column name of population.}

\item{magnitude}{a character (optional). A column name of magnitude. If \code{NULL}, magnitudes of all districts are replaced by \code{1}.}

\item{distinct_name}{If \code{TRUE}, all district names are distinct. Default is \code{FALSE}.}

\item{...}{ignored}
}
\value{
a \code{prcalc} object.
}
\description{
Transformation from \code{.csv} to \code{prcalc}
}
\details{
\code{.csv} file must include three columns---region(state) name, district name, and population (number of electorates). A column of magnitude is optional.
}
\examples{
\dontrun{
obj <- read_prcalc("data/my_file.csv")
                   region     = "region",
                   district   = "district",
                   population = "electorates",
                   magnitude  = "magnitude")
obj
}
}
\seealso{
\code{\link{as_prcalc}}
}
