% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prcalc.R
\name{read_prcalc}
\alias{read_prcalc}
\title{Transformation from \code{.csv} to \code{prcalc}}
\usage{
read_prcalc(x, l1, l2 = NULL, p, q = NULL, type = c("district", "party"), ...)
}
\arguments{
\item{x}{a \code{data.frame} or \code{tibble} object.}

\item{l1}{a character (mandatory). A column name of region or state.}

\item{l2}{a character (optional). A column name of party or district.}

\item{p}{a character (mandatory). A column name of population.}

\item{q}{a character (optional). A column name of magnitude. If \code{NULL}, magnitudes of all districts are replaced by \code{1}.}

\item{type}{a type of \code{l2}. If \code{"party"}, vector of \code{l2} represents parties. If \code{"district"}, vector of \code{l2} represents electoral district.}

\item{...}{ignored}
}
\value{
a \code{prcalc} object.
}
\description{
Transformation from \code{.csv} to \code{prcalc}
}
\details{
\code{.csv} file must include three columns---region(state) name, district name, and population (number of electorates). A column of magnitude is optional.
}
\examples{
\dontrun{
obj <- read_prcalc("data/my_file.csv",
                   l1   = "region",
                   l2   = "district",
                   p    = "electorates",
                   q    = "magnitude",
                   type = "district")
obj
}
}
\seealso{
\code{\link{as_prcalc}}
}
