% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prcalc.R
\name{read_prcalc}
\alias{read_prcalc}
\title{Transformation from \code{.csv} or \code{data.frame} to \code{prcalc_df}}
\usage{
read_prcalc(x, region = NULL, district = NULL, population = NULL, ...)
}
\arguments{
\item{x}{a data.frame or path of \code{.csv} file.}

\item{region}{a character. A column name of region or state.}

\item{district}{a character. A column name of district.}

\item{population}{a character. A column name of population.}

\item{...}{ignored

#' @details
\code{.csv} file must include three columns---region(state) name, district name, and population (number of electorates).}
}
\value{
a \code{prcalc_df} object.
}
\description{
Transformation from \code{.csv} or \code{data.frame} to \code{prcalc_df}
}
\examples{
\dontrun{
library(PRcalc)

my_df <- read_prcalc("data/my_file.csv")
my_df
}
}
