% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcalc.R
\name{summary.PRcalc}
\alias{summary.PRcalc}
\title{Summarise \code{PRclac} Object}
\usage{
\method{summary}{PRcalc}(object, prop = FALSE, digits = 3, ...)
}
\arguments{
\item{object}{\code{PRcalc} object}

\item{prop}{a logical value}

\item{digits}{a integer}

\item{...}{Ingnored}
}
\value{
A data.frame
}
\description{
Summarise \code{PRclac} Object
}
